% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query for MsigDB.}
\usage{
query(species = c("Hs", "Mm"), pathway = NULL, symbols = NULL, .unlist = FALSE)
}
\arguments{
\item{species}{species(Hs or Mm).}

\item{pathway}{regex expression to match possible pathsways.}

\item{symbols}{character vector of gene symbols to match possible pathsways}

\item{.unlist}{whether to unlist the symbols column.}
}
\description{
Query for MsigDB.
}
